Analysis

Visualization

using Plots
using QUBOTools

S = SampleSet([
    Sample([0, 0], 0.5,  8),
    Sample([0, 1], 1.2, 10),
    Sample([1, 0], 1.8, 12),
    Sample([1, 1], 1.5,  4),
])

plot(S)

Benchmarking

Timing

QUBOTools.total_time — Function

total_time(ω::SampleSet)

Retrieves the total time spent during the whole solution gathering process, as experienced by the user.

QUBOTools.effective_time — Function

effective_time(ω::SampleSet)

Retrieves the time spent by the algorithm in the strict sense, that is, excluding time spent with data access, precompilation and other activities. That said, it is assumed that $t_{\text{effective}} \le t_{\text{total}}$.

Solution Quality

QUBOTools.success_rate — Function

success_rate(ω::SampleSet{T,<:Any}, λ::T) where {T}

Returns the success rate according to the given sample set and the optimal objective value $\lambda$.

Time-to-Solution (TTS)

QUBOTools.tts — Function

tts(ω::SampleSet{T,<:Any}, λ::T, s::Float64=0.99) where {T}

Computes the time to solution (TTS) from the optimal objective value and a sample set. The success factor $s$ defaults to $0.99$.

tts(t::Float64, p::Float64, s::Float64=0.99)

Computes the time to solution (TTS) given the effective time $t$ spent running the algorithm and the success probability $p$. The success factor $s$ defaults to $0.99$.

\[\text{tts}(t, p; s) = t \frac{\log(1 - s)}{\log(1 - p)}\]