ToQUBO.jl Booklet

This booklet aims to gather the theoretical and practical details behind ToQUBO and provide documentation for project internals. The target audience includes, among others, advanced users and those willing to contribute to the project. The latter are advised to read the following sections, as they give a glimpse of the ideas employed up to now.

QUBO

\[\begin{array}{rl} \min & \mathbf{x}^{\intercal} Q\,\mathbf{x} \\ \text{s.t.} & \mathbf{x} \in \mathbb{B}^{n} \end{array}\]

ToQUBO.isqubo — Function

isqubo(model::MOI.ModelLike)

Tells if a given model is ready to be interpreted as a QUBO model.

For it to be true, a few conditions must be met:

All variables must be binary (MOI.VariableIndex ∈ MOI.ZeroOne)

No other constraints are allowed

The objective function must be of type MOI.ScalarQuadraticFunction, MOI.ScalarAffineFunction or MOI.VariableIndex

The objective sense must be either MOI.MIN_SENSE or MOI.MAX_SENSE

ToQUBO.toqubo — Function

toqubo(T::Type, source::MOI.ModelLike, optimizer::Union{Nothing, Type{<:MOI.AbstractOptimizer}} = nothing)
toqubo(source::MOI.ModelLike, optimizer::Union{Nothing, Type{<:MOI.AbstractOptimizer}} = nothing)

Low-level interface to create a ::VirtualQUBOModel{T} from ::MOI.ModelLike instance. If provided, an ::MOI.AbstractOptimizer is attached to the model.

ToQUBO.toqubo! — Function

toqubo!(model::VirtualQUBOModel{T}) where {T}

ToQUBO.toqubo_sense! — Function

toqubo_sense!(model::VirtualQUBOModel)

Copies MOI.ObjectiveSense from model.source_model to model.target_model.

ToQUBO.toqubo_variables! — Function

toqubo_variables!(model::VirtualQUBOModel{T}) where {T}

ToQUBO.toqubo_constraint! — Function

toqubo_constraint!(model::VirtualQUBOModel{T}, F::Type{<:SAF{T}}, S::Type{<:EQ{T}}) where {T}
toqubo_constraint!(model::VirtualQUBOModel{T}, F::Type{<:SAF{T}}, S::Type{<:LT{T}}) where {T}
toqubo_constraint!(model::VirtualQUBOModel{T}, F::Type{<:SQF{T}}, S::Type{<:EQ{T}}) where {T}
toqubo_constraint!(model::VirtualQUBOModel{T}, F::Type{<:SQF{T}}, S::Type{<:LT{T}}) where {T}
toqubo_constraint!(::VirtualQUBOModel{T}, ::Type{<:VI}, ::Type{<:Union{MOI.ZeroOne, MOI.Integer, MOI.Interval{T}, MOI.LessThan{T}, MOI.GreaterThan{T}}}) where {T}

ToQUBO.toqubo_objective! — Function

toqubo_objective!(model::VirtualQUBOModel{T}, F::Type{<:VI}) where {T}
toqubo_objective!(model::VirtualQUBOModel{T}, F::Type{<:SAF{T}}) where {T}
toqubo_objective!(model::VirtualQUBOModel{T}, F::Type{<:SQF{T}}) where {T}

Pseudo-Boolean Optimization

Internally, problems are represented through a Pseudo-Boolean Optimization (PBO) framework. The main goal is to represent a given problem using a Pseudo-Boolean Function (PBF) since there is an immediate correspondence between quadratic PBFs and QUBO forms.

ToQUBO.PBO.PseudoBooleanFunction — Type

PseudoBooleanFunction{S, T}(c::T)
PseudoBooleanFunction{S, T}(ps::Pair{Vector{S}, T}...)

A Pseudo-Boolean Function $f \in \mathscr{F}$ over some field $\mathbb{T}$ takes the form

\[f(\mathbf{x}) = \sum_{\omega \in \Omega\left[f\right]} c_\omega \prod_{j \in \omega} \mathbb{x}_j\]

where each $\Omega\left[{f}\right]$ is the multi-linear representation of $f$ as a set of terms. Each term is given by a unique set of indices $\omega \subseteq \mathbb{S}$ related to some coefficient $c_\omega \in \mathbb{T}$. We say that $\omega \in \Omega\left[{f}\right] \iff c_\omega \neq 0$. Variables $\mathbf{x}_i$ are indeed boolean, thus $f : \mathbb{B}^{n} \to \mathbb{T}$.

References

[1] Endre Boros, Peter L. Hammer, Pseudo-Boolean optimization, Discrete Applied Mathematics, 2002 {doi}

ToQUBO.PBO.residual — Function

residual(f::PBF{S, T}, i::S) where {S, T}
residual(f::PBF{S, T}, i::Int) where {S, T}

The residual of function $f \in \mathscr{F}$ with respect to the $i$-th variable.

\[ \Theta_i f(\mathbf{x}) = f(\mathbf{x}) - \mathbf{x}_i\, \Delta_i f(\mathbf{x}) = \sum_{\omega \in \Omega\left[{f}\right] \setminus \left\{{i}\right\}} c_{\omega} \prod_{k \in \omega} \mathbf{x}_k\]

ToQUBO.PBO.derivative — Function

derivative(f::PBF{S, T}, i::S) where {S, T}
derivative(f::PBF{S, T}, i::Int) where {S, T}

The partial derivate of function $f \in \mathscr{F}$ with respect to the $i$-th variable.

\[ \Delta_i f(\mathbf{x}) = \frac{\partial f(\mathbf{x})}{\partial \mathbf{x}_i} = \sum_{\omega \in \Omega\left[{f}\right] \setminus \left\{{i}\right\}} c_{\omega \cup \left\{{i}\right\}} \prod_{k \in \omega} \mathbf{x}_k\]

ToQUBO.PBO.gradient — Function

gradient(f::PBF)

Computes the gradient of $f \in \mathscr{F}$ where the $i$-th derivative is given by derivative.

ToQUBO.PBO.gap — Function

gap(f::PBF{S, T}; bound::Symbol=:loose) where {S, T}

Computes the least upper bound for the greatest variantion possible under some $f \in \mathscr{F}$ i. e.

\[\begin{array}{r l} \min & M \\ \text{s.t.} & \left|{f(\mathbf{x}) - f(\mathbf{y})}\right| \le M ~~ \forall \mathbf{x}, \mathbf{y} \in \mathbb{B}^{n} \end{array}\]

A simple approach, avaiable using the bound=:loose parameter, is to define

\[M \triangleq \sum_{\omega \neq \varnothing} \left|{c_\omega}\right|\]

ToQUBO.PBO.sharpness — Function

sharpness(f::PBF{S, T}; bound::Symbol=:loose, tol::T = T(1e-6)) where {S, T<:AbstractFloat}

ToQUBO.PBO.discretize — Function

discretize(f::PBF{S, T}; tol::T) where {S, T}

For a given function $f \in \mathscr{F}$ written as

\[ f\left({\mathbf{x}}\right) = \sum_{\omega \in \Omega\left[{f}\right]} c_\omega \prod_{i \in \omega} \mathbf{x}_i\]

computes an approximate function $g : \mathbb{B}^{n} \to \mathbb{Z}$ such that

\[ \argmin_{\mathbf{x} \in \mathbb{B}^{n}} g\left({\mathbf{x}}\right) = \argmin_{\mathbf{x} \in \mathbb{B}^{n}} f\left({\mathbf{x}}\right)\]

This is done by rationalizing every coefficient $c_\omega$ according to some tolerance tol.

ToQUBO.PBO.relaxed_gcd — Function

relaxed_gcd(x::T, y::T; tol::T = T(1e-6)) where {T <: AbstractFloat}

We define two real numbers $x$ and $y$ to be $\tau$-comensurable if, for some $\tau > 0$ there exists a continued fractions convergent $p_{k} \div q_{k}$ such that

\[ \left| {q_{k} x - p_{k} y} \right| \le \tau\]

Quadratization

In order to successfully achieve a QUBO formulation, sometimes it is needed to quadratize the resulting PBF, i.e., reduce its degree until reaching the quadratic case. There are many quadratization methods available, and ToQUBO implements a few of them.

ToQUBO.PBO.quadratize — Function

quadratize(f::PBF{S, T}; slack::Any) where {S, T}

Quadratizes a given PBF, i.e. creates a function $g \in \mathscr{F}^{2}$ from $f \in \mathscr{F}^{k}, k \ge 3$.

A function $f : 2^{S} \to \mathbb{R}$ is said to be submodular if

\[f(X \cup Y) + f(X \cap Y) \le f(X) + f(Y) \forall X, Y \subset S\]

ToQUBO.PBO.@quadratization — Macro

@quadratization(name, nsv, nst)

Defines a new quadratization technique.

Virtual Mapping

During reformulation, ToQUBO holds two distinct models, namely the Source Model and the Target Model. The source model is a generic MOI model restricted to the supported constraints. The target one is on the QUBO form used during the solving process. Both lie within a Virtual Model, which provides the necessary API integration and keeps all variable and constraint mapping tied together.

This is done in a transparent fashion for both agents since the user will mostly interact with the presented model, and the solvers will only access the generated one.

Virtual Variables

Every virtual model stores a collection of virtual variables, intended to provide a link between those in the source and those to be created in the target model. Each virtual variable stores enconding information for later expansion and evaluation.

ToQUBO.VirtualMapping.VirtualVariable — Type

VirtualVariable{S, T}(
    newvar::Function,
    source::Union{S, Nothing};
    bits::Union{Int, Nothing},
    tech::Symbol,
    name::Union{Symbol, Nothing}=nothing,
    α::T=zero(T),
    β::T=one(T)
) where {S, T}

The Virtual Variable Mapping

Variable Expansion techniques:

:𝔹 - Used when a boolean variable is to be mirrored.

:ℤ₂ - Binary expansion for integer variable.

:ℤ₁ - Unary expansion for integer variable.

:ℝ₂ - Binary expansion for real variable.

:ℝ₁ - Unary expansion for real variable.

References:

[1] Chancellor, N. (2019). Domain wall encoding of discrete variables for quantum annealing and QAOA. Quantum Science and Technology, 4(4), 045004. {doi}

ToQUBO.VirtualMapping.mapvar! — Function

mapvar!(model::AbstractVirtualModel{T}, v::VirtualMOIVariable{T}) where {T}

Maps newly created virtual variable v within the virtual model structure. It follows these steps:

Maps v's source to it in the model's source mapping.

For every one of v's targets, maps it to itself and adds a binary constraint to it.

Adds v to the end of the model's varvec.

ToQUBO.VirtualMapping.expandℝ! — Function

expandℝ!(model::QUBOModel{T}, src::VI; bits::Int, name::Symbol, α::T, β::T, semi::Bool) where T

Real Binary Expansion within the closed interval $[\alpha, \beta]$.

For a given variable $x \in [\alpha, \beta]$ we approximate it by

\[x \approx \alpha + \frac{(\beta - \alpha)}{2^{n} - 1} \sum_{i=0}^{n-1} {2^{i}\, y_i}\]

where $n$ is the number of bits and $y_i \in \mathbb{B}$.

ToQUBO.VirtualMapping.slackℝ! — Function

slackℝ!(model::AbstractVirtualModel{T}; name::Symbol, α::T, β::T, semi::Bool) where T

Adds real slack variable according to expandℝ!'s expansion method.

ToQUBO.VirtualMapping.expandℤ! — Function

expandℤ!(model::QUBOModel{T}, src::VI; name::Symbol, α::T, β::T, semi::Bool) where T

Integer Binary Expansion within the closed interval $[\left\lceil{\alpha}\right\rceil, \left\lfloor{\beta}\right\rfloor]$.

ToQUBO.VirtualMapping.slackℤ! — Function

slackℤ!(model::AbstractVirtualModel{T}; name::Symbol, α::T, β::T) where {T}

Adds integer slack variable according to expandℤ!'s expansion method.

ToQUBO.VirtualMapping.mirror𝔹! — Function

mirror𝔹!(model::AbstractVirtualModel{T}, src::Union{VI, Nothing}; name::Symbol) where T

Simply crates a virtual-mapped Doppelgänger into the destination model.

ToQUBO.VirtualMapping.slack𝔹! — Function

slack𝔹!(model::AbstractVirtualModel{T}; name::Symbol) where {T}

Adds a binary slack variable to the model.

Virtual Models

ToQUBO.VirtualMapping.AbstractVirtualModel — Type

abstract type AbstractVirtualModel{T} <: MOI.AbstractOptimizer end

ToQUBO.VirtualQUBOModel — Type

VirtualQUBOModel{T}(optimizer::Union{Nothing, Type{<:MOI.AbstractOptimizer}} = nothing) where {T}

This QUBO Virtual Model links the final QUBO formulation to the original one, allowing variable value retrieving and other features.

Annealing & Sampling

ToQUBO's main goal is to benefit from non-deterministic samplers, especially Quantum Adiabatic devices and other Annealing machines. A few MOI-compliant interfaces for annealers and samplers are bundled within ToQUBO via the Anneal.jl submodule and package prototype. Some of them are presented below.

Quantum Annealing

Interfacing with D-Wave's quantum computers is one of the milestones we expect to achieve with this package. Like other proprietary optimization resources such as Gurobi and FICO® Xpress, this requires licensing and extra steps are needed to get access to it. In a first moment, for those willing to get started, the Simulated Annealing optimizer might be enough.

While in JuMP, run using Anneal and look for QuantumAnnealer.Optimizer.

Simulated Annealing

Provided by D-Wave's open-source code libraries, this Simulated Annealing engine implements some of the features and configuration you would find using the Quantum API. Its adoption is recommended for basic usage, tests, and during early research steps due to its simplicity and ease of use. It does not implement the most advanced Simulated Annealing algorithm available but performs fairly well on small instances. Anneal.jl exports this interface as SimulatedAnnealer.Optimizer.

Random Sampling

This sampler is implemented for test purposes and simply assigns 0 or 1 to each variable according to a given probability bias $0 \le p \le 1$, which defaults to $p = 0.5$. After running the using Anneal command, RandomSampler.Optimizer will be available.

Exact Solver (Exaustive Enumeration)

Also made to be used in tests, the ExactSolver.Optimizer interface runs through all possible state configurations, which implies in an exponential time complexity on the number of variables. Thus, only problems with no more than 20 variables should be provided.

MIQP Solvers

The most accessible alternative to Annealers and Samplers are Mixed-Integer Quadratic Programming (MIQP) Solvers such as Gurobi and CPLEX. These are not intended to be of regular use alongside ToQUBO since reformulation usually makes things harder for these folks. Yet, there are still cases where they may be suitable for tests on small instances.

Custom Error Types

ToQUBO.QUBOError — Type

QUBOError(msg::Union{Nothing, String})

This error indicates any failure during QUBO formulation