Pseudo-Boolean Optimization

Internally, problems are represented through a Pseudo-Boolean Optimization (PBO) framework. The main goal is to represent a given problem using a Pseudo-Boolean Function (PBF) since there is an immediate correspondence between optimization over quadratic PBFs and the QUBO formalism.

ToQUBO.PBO.PseudoBooleanFunction — Type

PseudoBooleanFunction{V,T}(Ω::Dict{Union{Set{V},Nothing},T}) where {V,T}

A Pseudo-Boolean Function[Boros2002]$f \in \mathscr{F}$ over some field $\mathbb{T}$ takes the form

\[f(\mathbf{x}) = \sum_{\omega \in \Omega\left[f\right]} c_\omega \prod_{j \in \omega} x_j\]

where each $\Omega\left[{f}\right]$ is the multi-linear representation of $f$ as a set of terms. Each term is given by a unique set of indices $\omega \subseteq \mathbb{S}$ related to some coefficient $c_\omega \in \mathbb{T}$. We say that $\omega \in \Omega\left[{f}\right] \iff c_\omega \neq 0$. Variables $x_j$ are boolean, thus $f : \mathbb{B}^{n} \to \mathbb{T}$.

Quadratization

In order to successfully achieve a QUBO formulation, sometimes it is needed to quadratize the resulting PBF, i.e., reduce its degree until reaching the quadratic case.

A quadratization is a mapping $\mathcal{Q}: \mathscr{F} \to \mathscr{F}^{2}$ such that

\[\forall f \in \mathscr{F}, \forall x \in \{0, 1\}^{n}, \min_{y} \mathcal{Q}\left\lbrace{}f\right\rbrace{}(x; y) = f(x) \]

There are many quadratization methods available[Dattani2019], and ToQUBO implements two of them for now. However, using Julia's multiple dispatch paradigm, it's possible to extend the quadratization method coverage with your own algorithms.

ToQUBO.PBO.quadratize! — Function

quadratize!(aux::Function, f::PBF{S, T}, ::Quadratization{Q}) where {S,T,Q}

Quadratizes a given PBF in-place, i.e. applies a mapping $Q : \mathscr{F}^{k} \to \mathscr{F}^{2}$.

aux(::Nothing)::S
aux(::Integer)::Vector{S}

Implemented Quadratization Techniques

Currently, ToQUBO has two reduction algorithms, one for negative and another for positive terms.

ToQUBO.PBO.NTR_KZFD — Type

Quadratization{NTR_KZFD}(stable::Bool = false)

NTR-KZFD (Kolmogorov & Zabih, 2004; Freedman & Drineas, 2005)

Info

Introduces one new variable and no non-submodular terms.

ToQUBO.PBO.PTR_BG — Type

Quadratization{PTR_BG}(stable::Bool = false)

PTR-BG (Boros & Gruber, 2014)

Info

Introduces $k - 2$ new variables and $k - 1$ non-submodular terms.

A Primer on Submodularity

A set function $f : 2^{S} \to \mathbb{R}$ is said to be submodular if

\[f(X \cup Y) + f(X \cap Y) \le f(X) + f(Y) \forall X, Y \subset S\]

holds.

Boros2002Endre Boros, Peter L. Hammer, Pseudo-Boolean optimization, Discrete Applied Mathematics, 2002 {doi}

Dattani2019Nikesh S. Dattani, Quadratization in discrete optimization and quantum mechanics, ArXiv, 2019 {doi}