Manual

Quick Start Guide

using JuMP
using ToQUBO
using DWaveNeal # <- Your favourite Annealer/Sampler/Solver here

model = Model(() -> ToQUBO.Optimizer(DWaveNeal.Optimizer))

@variable(model, x[1:3], Bin)

@objective(model, Max, 1.0 * x[1] + 2.0 * x[2] + 3.0 * x[3])

@constraint(model, 0.3 * x[1] + 0.5 * x[2] + 1.0 * x[3] <= 1.6)

optimize!(model)

solution_summary(model)

* Solver : Virtual Model

* Status
  Result count       : 21
  Termination status : LOCALLY_SOLVED
  Message from the solver:
  ""

* Candidate solution (result #1)
  Primal status      : FEASIBLE_POINT
  Dual status        : NO_SOLUTION
  Objective value    : 5.00000e+00

* Work counters
  Solve time (sec)   : 2.37788e-01

Compiler Flags

Architecture

ToQUBO.Attributes.Architecture — Type

Architecture()

Quadratization

ToQUBO.Attributes.Quadratize — Type

Quadratize()

Boolean flag to conditionally perform the quadratization step. Is automatically set by the compiler when high-order functions are generated.

ToQUBO.Attributes.QuadratizationMethod — Type

QuadratizationMethod()

Defines which quadratization method to use. Available options are defined in the PBO submodule.

ToQUBO.Attributes.StableQuadratization — Type

StableQuadratization()

When set, this boolean flag enables stable quadratization methods, thus yielding predictable results. This is intended to be used during tests or other situations where deterministic output is desired. On the other hand, usage in production is not recommended since it requires increased memory and processing resources.

Variable & Constraint Encoding

ToQUBO.Attributes.VariableEncodingMethod — Type

VariableEncodingMethod()

Available methods are:

Binary (default)

Unary

Arithmetic

OneHot

DomainWall

Bounded

The Binary, Unary and Arithmetic encodings can have their expansion coefficients bounded by parametrizing the Bounded encoding.

ToQUBO.Attributes.VariableEncodingPenalty — Type

VariableEncodingPenalty()

Allows the user to set and retrieve the coefficients used for encoding variables when additional constraints are involved.

ToQUBO.Attributes.ConstraintEncodingPenalty — Type

ConstraintEncodingPenalty()

Allows the user to set and retrieve the coefficients used for encoding constraints.