API Reference

Fallback dispatch

When extending QUBOTools, one must implement a method for QUBOTools.backend.

QUBOTools.backend — Function

backend(model)::AbstractModel
backend(model::AbstractModel)::AbstractModel

Retrieves the model's backend. Implementing this function allows one to profit from fallback implementations of the other methods.

Variable System

QUBOTools.varlt — Function

varlt(x::V, y::V) where {V}

This function exists to define an arbitrary ordering for a given type and was created to address [1]. There is no predefined comparison between instances MOI's VariableIndex type. [1] https://github.com/jump-dev/MathOptInterface.jl/issues/1985

Variable Domains

QUBOTools.Domain — Type

Domain

QUBOTools.BoolDomain — Type

BoolDomain <: Domain

\[x \in \mathbb{B} = \lbrace{0, 1}\rbrace\]

QUBOTools.SpinDomain — Type

SpinDomain <: Domain

\[s \in \mathbb{S} = \lbrace{-1, 1}\rbrace\]

Missing docstring.

Missing docstring for QUBOTools.UnknownDomain. Check Documenter's build log for details.

QUBOTools.domain — Function

domain(model::AbstractModel)::Domain
domain(fmt::AbstractFormat)::Domain

Returns the singleton representing the variable domain of a given model.

QUBOTools.sense — Function

sense(model)::Sense

Frame Casting

QUBOTools.cast — Function

cast(
    source_sense::Sense,
    source_domain::Domain,
    target_sense::Sense,
    target_domain::Domain,
    x::Any
)

cast(::S, ::S, model::AbstractModel) where {S<:Sense}
cast(::S1, ::S2, model::AbstractModel) where {S1<:Sense,S2<:Sense}

Recasting the sense of a model preserves its meaning:

\[\begin{array}{ll} \min_{s} \alpha [f(s) + \beta] &\equiv \max_{s} -\alpha [f(s) + \beta] \\ &\equiv \max_{s} \alpha [-f(s) - \beta] \\ \end{array}\]

The linear terms, quadratic terms and constant offset of a model have its signs reversed.

cast(::S, ::S, s::Sample) where {S<:Sense}
cast(::Sense, ::Sense, s::Sample) where {S1<:Sense,S2<:Sense}
cast(::S, ::S, ω::SampleSet) where {S<:Sense}
cast(::Sense, ::Sense, ω::SampleSet) where {S1<:Sense,S2<:Sense}

cast(target, model::AbstractModel)
cast(source, target, ψ::Vector{U})
cast(source, target, Ψ::Vector{Vector{U}})
cast(source, target, ω::SampleSet)

Returns a new object, switching its domain from source to target.

Reverses the sign of the objective value.

Solution Interface

QUBOTools.Sample — Type

Sample{T,U}(state::Vector{U}, value::T, reads::Integer) where{T,U}

QUBOTools.SampleSet — Type

SampleSet{T,U}(
    data::Vector{Sample{T,U}},
    metadata::Dict{String, Any},
) where {T,U}

It compresses repeated states by adding up the reads field. It was inspired by [1], with a few tweaks.

Info

A SampleSet{T,U} was designed to be read-only. It is optimized to support queries over the solution set.

References

[1] ocean docs

QUBOTools.sampleset — Function

sampleset(model)::SampleSet

Returns the SampleSet stored in a model.

QUBOTools.state — Function

state(model, i::Integer)

Returns the state vector corresponding to the $i$-th solution on the model's sampleset.

QUBOTools.value — Function

value(model, state::Vector)
value(model, index::Integer)

This function aims to evaluate the energy of a given state under some QUBO Model.

value(Q::Dict{Tuple{Int,Int},T}, ψ::Vector{U}, α::T = one(T), β::T = zero(T)) where {T}
value(h::Dict{Int,T}, J::Dict{Tuple{Int,Int},T}, ψ::Vector{U}, α::T = one(T), β::T = zero(T)) where {T}

Info

Scale and offset factors are taken into account.

QUBOTools.energy — Function

energy

An alias for value.

QUBOTools.reads — Function

reads(model)
reads(model, i::Integer)

Returns the read frequency of the $i$-th solution on the model's sampleset.

Models

QUBOTools.AbstractModel — Type

AbstractModel{V,T}

Represents an abstract QUBO Model and should support most of the queries made available by QUBOTools. ```

As shown in the example above, implementing a method for the backend function gives access to most fallback implementations.

QUBOTools.Model — Type

Model{
    V <: Any,
    T <: Real,
    U <: Integer
} <: AbstractModel{V,T}

The Model was designed to work as a general for all implemented interfaces. It is intended to be the core engine behind the target codecs.

MathOptInterface/JuMP

Both V <: Any and T <: Real parameters exist to support MathOptInterface/JuMP integration. By choosing V = MOI.VariableIndex and T matching Optimizer{T} the hard work should be done.

QUBOTools.model_name — Function

model_name(model)::String

Returns a string representing the model type.

QUBOTools.AbstractFormat — Type

AbstractFormat

QUBOTools.infer_format — Function

infer_format(::AbstractString)::AbstractFormat
infer_format(::Symbol)::AbstractFormat
infer_format(::Symbol, ::Symbol)::AbstractFormat

Given a file path, tries to infer the type associated to a QUBO model format.

Data Access

QUBOTools.linear_terms — Function

linear_terms(model)::Dict{Int,T} where {T <: Real}

Retrieves the linear terms of a model as a dict.

Info

The explicit_linear_terms method includes all variables, breaking linear sparsity.

QUBOTools.explicit_linear_terms — Function

explicit_linear_terms(model)::Dict{Int,T} where {T <: Real}

Retrieves the linear terms of a model as a dict, including zero entries.

QUBOTools.quadratic_terms — Function

quadratic_terms(model)::Dict{Tuple{Int,Int},T} where {T <: Real}

Retrieves the quadratic terms of a model as a dict. For every key pair $(i, j)$ holds that $i < j$.

QUBOTools.scale — Function

scale(model)

QUBOTools.offset — Function

offset(model)

QUBOTools.variable_map — Function

variable_map(model)::Dict{V,Int} where {V}
variable_map(model, x::V)::Integer where {V}

QUBOTools.variable_inv — Function

variableinv(model)::Dict{Int,V} where {V} variableinv(model, i::Integer)::V where {V}

QUBOTools.variable_set — Function

variable_set(model)::Set

Returns the set of variables of a given model.

QUBOTools.variables — Function

variables(model)::Vector

Returns a sorted vector containing the model's variables. If order doesn't matter, use variable_set(model) instead.

QUBOTools.id — Function

id(model)

QUBOTools.version — Function

version(model)

QUBOTools.description — Function

description(model)

QUBOTools.metadata — Function

metadata(model)

Queries

QUBOTools.domain_size — Function

domain_size(model)::Integer

Counts the number of variables in the model.

QUBOTools.linear_size — Function

linear_size(model)::Int

Counts the number of non-zero linear terms in the model.

QUBOTools.quadratic_size — Function

quadratic_size(model)::Int

Counts the number of non-zero quadratic terms in the model.

QUBOTools.density — Function

density(model)::Float64

Computes the density $\rho$ of non-zero terms in a model, according to the expression

\[\rho = \frac{2Q + L}{N^{2}}\]

where $L$ is the number of non-zero linear terms, $Q$ the number of quadratic ones and $N$ the number of variables.

If the model is empty, returns NaN.

QUBOTools.linear_density — Function

linear_density(model)::Float64

Computes the linear density $\rho_{l}$, given by

\[\rho_{l} = \frac{L}{N}\]

where $L$ is the number of non-zero linear terms and $N$ the number of variables.

QUBOTools.quadratic_density — Function

quadratic_density(model)::Float64

Computes the linear density $\rho_{q}$, given by

\[\rho_{q} = \frac{2Q}{N (N - 1)}\]

where $Q$ is the number of non-zero quadratic terms and $N$ the number of variables.

QUBOTools.adjacency — Function

adjacency(model)::Dict{Int,Set{Int}}
adjacency(G::Vector{Tuple{Int,Int}})::Dict{Int,Set{Int}}
adjacency(G::Set{Tuple{Int,Int}})::Dict{Int,Set{Int}}
adjacency(G::Dict{Tuple{Int,Int},T})::Dict{Int,Set{Int}}

Computes the adjacency list representation for the quadratic terms of a given model. A mapping between each variable index and the set of its neighbors is returned.

adjacency(model, k::Integer)::Set{Int}
adjacency(G::Vector{Tuple{Int,Int}}, k::Integer)::Set{Int}
adjacency(G::Set{Tuple{Int,Int}}, k::Integer)::Set{Int}
adjacency(G::Dict{Tuple{Int,Int},T}, k::Integer)::Set{Int}

If a second parameter, an integer, is present, then the set of neighbors of that node is returned.

Warning

Computing specific neighborhoods is expensive. Thus, it is recommended that one stores the adjacency list for repeated access.

Normal Forms

QUBOTools.qubo — Function

qubo(model::AbstractModel{<:BoolDomain})
qubo(model::AbstractModel{<:BoolDomain}, ::Type{Dict}, T::Type = Float64)

Returns sparse dictionary representation.

qubo(model::AbstractModel{<:BoolDomain}, ::Type{Vector}, T::Type = Float64)

Returns sparse vector quadruple (linear, quadratic, lower index & upper index).

qubo(model::AbstractModel{<:BoolDomain}, ::Type{Matrix}, T::Type = Float64)

Returns dense matrix representation.

qubo(model::AbstractModel{<:BoolDomain}, ::Type{SparseMatrixCSC}, T::Type = Float64)

Returns sparse matrix representation.

qubo(model::AbstractModel{<:SpinDomain}, args...)

Returns QUBO form from Ising Model (Spin).

qubo(h::Dict{Int,T}, J::Dict{Tuple{Int, Int}, T}, α::T = one(T), β::T = zero(T)) where {T}    
qubo(h::Vector{T}, J::Vector{T}, u::Vector{Int}, v::Vector{Int}, α::T = one(T), β::T = zero(T)) where {T}
qubo(h::Vector{T}, J::Matrix{T}, α::T = one(T), β::T = zero(T)) where {T}
qubo(h::SparseVector{T}, J::SparseMatrixCSC{T}, α::T = one(T), β::T = zero(T)) where {T}

Info

Apart from the sparse matricial case, the linear terms are explicitly included, breaking sparsity by containing zero entries.

QUBOTools.ising — Function

ising(model::AbstractModel{<:SpinDomain})
ising(model::AbstractModel{<:SpinDomain}, ::Type{<:Dict}, T::Type = Float64))

Returns sparce dictionary representation

ising(model::AbstractModel{<:BoolDomain}, ::Type{Vector}, T::Type = Float64)

Returns sparse vector quadruple (linear, quadratic, lower index & upper index).

ising(model::AbstractModel{<:BoolDomain}, ::Type{Matrix}, T::Type = Float64)

Returns dense matrix representation.

ising(model::AbstractModel{<:SpinDomain}, ::Type{SparseMatrixCSC}, T::Type = Float64)

Returns sparce matrix representation

ising(model::AbstractModel{<:BoolDomain}, args...)

Returns Ising Model form from QUBO Model (Bool).

ising(h::Dict{Int,T}, J::Dict{Tuple{Int, Int}, T}, α::T = one(T), β::T = zero(T)) where {T}    
ising(h::Vector{T}, J::Vector{T}, u::Vector{Int}, v::Vector{Int}, α::T = one(T), β::T = zero(T)) where {T}
ising(h::Vector{T}, J::Matrix{T}, α::T = one(T), β::T = zero(T)) where {T}
ising(h::SparseVector{T}, J::SparseMatrixCSC{T}, α::T = one(T), β::T = zero(T)) where {T}

Ising Normal Form

Info

Apart from the sparse matricial case, the linear terms are explicitly included, breaking sparsity by containing zero entries.