ToQUBO.jl Booklet

This booklet aims to gather the theoretical and practical details behind ToQUBO and provide documentation for project internals. The target audience includes, among others, advanced users and those willing to contribute to the project. The latter are advised to read the following sections, as they give a glimpse of the ideas employed up to now.

QUBO

\[\begin{array}{rl} \min & \mathbf{x}^{\intercal} Q\,\mathbf{x} \\ \text{s.t.} & \mathbf{x} \in \mathbb{B}^{n} \end{array}\]

ToQUBO.isqubo — Function

isqubo(model::MOI.ModelLike)

Tells if a given model is ready to be interpreted as a QUBO model.

For it to be true, a few conditions must be met:

All variables must be binary (MOI.VariableIndex ∈ MOI.ZeroOne)

No other constraints are allowed

The objective function must be of type MOI.ScalarQuadraticFunction, MOI.ScalarAffineFunction or MOI.VariableIndex

The objective sense must be either MOI.MIN_SENSE or MOI.MAX_SENSE

ToQUBO.toqubo — Function

toqubo(
    [T=Float64,]
    source::MOI.ModelLike,
    ::AbstractArchitecture;
    optimizer::Union{Nothing, Type{<:MOI.AbstractOptimizer}} = nothing
)

Low-level interface to create a ::VirtualModel{T} from ::MOI.ModelLike instance. If provided, an ::MOI.AbstractOptimizer is attached to the model.

ToQUBO.toqubo! — Function

toqubo!(model::VirtualModel{T}, ::AbstractArchitecture) where {T}

ToQUBO.toqubo_sense! — Function

toqubo_sense!(model::VirtualModel, ::AbstractArchitecture) where {T}

Copies MOI.ObjectiveSense from model.source_model to model.target_model.

ToQUBO.toqubo_variables! — Function

toqubo_variables!(model::VirtualModel{T}) where {T}

ToQUBO.toqubo_constraint — Function

toqubo_constraint

Returns the pseudo-boolean function associated to a given constraint from the source model.

ToQUBO.toqubo_constraints! — Function

toqubo_constraints!(model::VirtualModel, ::AbstractArchitecture)

ToQUBO.toqubo_objective! — Function

toqubo_objective!(model::VirtualModel, ::AbstractArchitecture)

ToQUBO.toqubo_penalties! — Function

toqubo_penalties!(model::VirtualModel, ::AbstractArchitecture)

ToQUBO.toqubo_parse! — Function

toqubo_parse!(
    model::VirtualModel{T},
    g::PBO.PBF{VI,T},
    f::MOI.AbstractFunction,
    arch::AbstractArchitectur
) where {T}

Parses the given MOI function f into PBF g.

ToQUBO.toqubo_build! — Function

toqubo_build!(model::VirtualModel, ::AbstractArchitecture)

Pseudo-Boolean Optimization

Internally, problems are represented through a Pseudo-Boolean Optimization (PBO) framework. The main goal is to represent a given problem using a Pseudo-Boolean Function (PBF) since there is an immediate correspondence between quadratic PBFs and QUBO forms.

ToQUBO.PBO.PseudoBooleanFunction — Type

PseudoBooleanFunction{V,T}(Ω::Dict{Union{Set{V},Nothing},T}) where {V,T}

A Pseudo-Boolean Function[Boros2002]$f \in \mathscr{F}$ over some field $\mathbb{T}$ takes the form

\[f(\mathbf{x}) = \sum_{\omega \in \Omega\left[f\right]} c_\omega \prod_{j \in \omega} x_j\]

where each $\Omega\left[{f}\right]$ is the multi-linear representation of $f$ as a set of terms. Each term is given by a unique set of indices $\omega \subseteq \mathbb{S}$ related to some coefficient $c_\omega \in \mathbb{T}$. We say that $\omega \in \Omega\left[{f}\right] \iff c_\omega \neq 0$. Variables $x_j$ are boolean, thus $f : \mathbb{B}^{n} \to \mathbb{T}$.

Missing docstring.

Missing docstring for ToQUBO.PBO.residual. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ToQUBO.PBO.derivative. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ToQUBO.PBO.gradient. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ToQUBO.PBO.gap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ToQUBO.PBO.sharpness. Check Documenter's build log for details.

ToQUBO.PBO.relaxed_gcd — Function

relaxed_gcd(x::T, y::T; tol::T = T(1e-6)) where {T}

We define two real numbers $x$ and $y$ to be $\tau$-comensurable if, for some $\tau > 0$ there exists a continued fractions convergent $p_{k} \div q_{k}$ such that

\[ \left| {q_{k} x - p_{k} y} \right| \le \tau\]

A Primer on Submodularity

A set function $f : 2^{S} \to \mathbb{R}$ is said to be submodular if

\[f(X \cup Y) + f(X \cap Y) \le f(X) + f(Y) \forall X, Y \subset S\]

holds.

Quadratization

In order to successfully achieve a QUBO formulation, sometimes it is needed to quadratize the resulting PBF, i.e., reduce its degree until reaching the quadratic case. There are many quadratization methods available, and ToQUBO implements a few of them.

ToQUBO.PBO.quadratize! — Function

quadratize!(aux::Function, f::PBF{S, T}, ::Quadratization{Q}) where {S,T,Q}

Quadratizes a given PBF in-place, i.e. applies a mapping $Q : \mathscr{F}^{k} \to \mathscr{F}^{2}$.

aux(::Nothing)::S
aux(::Integer)::Vector{S}

Virtual Mapping

During reformulation, ToQUBO holds two distinct models, namely the Source Model and the Target Model. The source model is a generic MOI model restricted to the supported constraints. The target one is on the QUBO form used during the solving process. Both lie within a Virtual Model, which provides the necessary API integration and keeps all variable and constraint mapping tied together.

This is done in a transparent fashion for both agents since the user will mostly interact with the presented model, and the solvers will only access the generated one.

Virtual Variables

Every virtual model stores a collection of virtual variables, intended to provide a link between those in the source and those to be created in the target model. Each virtual variable stores enconding information for later expansion and evaluation.

ToQUBO.VirtualVariable — Type

Variable Expansion methods:

- Linear
- Unary
- Binary
- One Hot
- Domain Wall

References:

[1] Chancellor, N. (2019). Domain wall encoding of discrete variables for quantum annealing and QAOA. Quantum Science and Technology, 4(4), 045004. {doi}

ToQUBO.encode! — Function

encode!(model::VirtualModel{T}, v::VirtualVariable{T}) where {T}

Maps newly created virtual variable v within the virtual model structure. It follows these steps:

Maps v's source to it in the model's source mapping.

For every one of v's targets, maps it to itself and adds a binary constraint to it.

Adds v to the end of the model's varvec.

Variable Encoding

Linear Encoding Methods

ToQUBO.LinearEncoding — Type

LinearEncoding

Every linear encoding $\xi$ is of the form

\[\xi(\mathbf{y}) = \alpha + \sum_{i = 1}^{n} \gamma_{i} y_{i}\]

ToQUBO.Binary — Type

Binary

Binary Expansion within the closed interval $[\alpha, \beta]$.

For a given variable $x \in [\alpha, \beta]$ we approximate it by

\[x \approx \alpha + \frac{(\beta - \alpha)}{2^{n} - 1} \sum_{i=0}^{n-1} {2^{i}\, y_i}\]

where $n$ is the number of bits and $y_i \in \mathbb{B}$.

ToQUBO.Unary — Type

Unary()

Let $x \in [a, b] \subset \mathbb{Z}, n = b - a, \mathbf{y} \in \mathbb{B}^{n}$.

\[x \approx \xi(\mathbf{y}) = a + \sum_{j = 1}^{b - a} y_{j}\]

ToQUBO.Arithmetic — Type

Arithmetic

ToQUBO.OneHot — Type

OneHot()

The one-hot encoding is a linear technique used to represent a variable $x \in \left\lbrace{\gamma_{j}}}\right\rbrace_{j \in \left[n\right]$.

The encoding function is combined with a constraint assuring that only one and exactly one of the expansion's variables $y_{j}$ is activated at a time.

\[\begin{array}{rl} x = \xi(\mathbf{y}) = & \sum_{j = 1}^{n} \gamma_{j} y_{j} \\ \mathrm{s.t.} & \sum_{j = 1}^{n} y_{j} = 1 \end{array}\]

Sequential Encoding Methods

Missing docstring.

Missing docstring for ToQUBO.SequentialEncoding. Check Documenter's build log for details.

ToQUBO.DomainWall — Type

DomainWall()

The Domain Wall[Chancellor2019] encoding method is a sequential approach that requires only $n - 1$ bits to represent $n$ distinct values.

Encoding Analysis

bits	linear	quadratic	$\Delta$

Domain Wall	$n - 1$	$n$		$O(n)$

Bounded Coefficients

ToQUBO.Bounded — Type

Bounded{Binary,T}(μ::T) where {T}

Let $x \in [a, b] \subset \mathbb{Z}$ and $n = b - a$.

First,

\[\begin{align*} 2^{k - 1} &\le \mu \\ \implies k &= \left\lfloor\log_{2} \mu + 1 \right\rfloor \end{align*}\]

Since

\[\sum_{j = 1}^{k} 2^{j - 1} = \sum_{j = 0}^{k - 1} 2^{j} = 2^{k} - 1\]

then, for $r \in \mathbb{N}$

\[n = 2^{k} - 1 + r \times \mu + \epsilon \implies r = \left\lfloor \frac{n - 2^{k} + 1}{\mu} \right\rfloor\]

and

\[\epsilon = n - 2^{k} + 1 - r \times \mu\]

\gamma_{j} = \left\lbrace\begin{array}{cl} 2^{j} & \text{if } 1 \le j \le k \ \mu & \text{if } k < j < r + k \ n - 2^k + 1 - r \times \mu & \text{otherwise} \end{array}\right. ```

Bounded{Unary,T}(μ::T) where {T}

Virtual Models

ToQUBO.VirtualModel — Type

VirtualModel{T}(optimizer::Union{Nothing, Type{<:MOI.AbstractOptimizer}} = nothing) where {T}

This Virtual Model links the final QUBO formulation to the original one, allowing variable value retrieving and other features.

Annealing & Sampling

ToQUBO's main goal is to benefit from non-deterministic samplers, especially Quantum Adiabatic devices and other Annealing machines. A few MOI-compliant interfaces for annealers and samplers are bundled within ToQUBO via the Anneal.jl submodule and package prototype. Some of them are presented below.

Quantum Annealing

Interfacing with D-Wave's quantum computers is one of the milestones we expect to achieve with this package. Like other proprietary optimization resources such as Gurobi and FICO® Xpress, this requires licensing and extra steps are needed to get access to it. In a first moment, for those willing to get started, the Simulated Annealing optimizer might be enough.

While in JuMP, run using Anneal and look for QuantumAnnealer.Optimizer.

Simulated Annealing

Provided by D-Wave's open-source code libraries, this Simulated Annealing engine implements some of the features and configuration you would find using the Quantum API. Its adoption is recommended for basic usage, tests, and during early research steps due to its simplicity and ease of use. It does not implement the most advanced Simulated Annealing algorithm available but performs fairly well on small instances. Anneal.jl exports this interface as SimulatedAnnealer.Optimizer.

Random Sampling

This sampler is implemented for test purposes and simply assigns 0 or 1 to each variable according to a given probability bias $0 \le p \le 1$, which defaults to $p = 0.5$. After running the using Anneal command, RandomSampler.Optimizer will be available.

Exact Solver (Exaustive Enumeration)

Also made to be used in tests, the ExactSolver.Optimizer interface runs through all possible state configurations, which implies in an exponential time complexity on the number of variables. Thus, only problems with no more than 20 variables should be provided.

MIQP Solvers

The most accessible alternative to Annealers and Samplers are Mixed-Integer Quadratic Programming (MIQP) Solvers such as Gurobi and CPLEX. These are not intended to be of regular use alongside ToQUBO since reformulation usually makes things harder for these folks. Yet, there are still cases where they may be suitable for tests on small instances.

Custom Error Types

ToQUBO.QUBOError — Type

QUBOError(msg::Union{Nothing, String})

This error indicates any failure during QUBO formulation

Boros2002Endre Boros, Peter L. Hammer, Pseudo-Boolean optimization, Discrete Applied Mathematics, 2002 {doi}

Chancellor2019Nicholas Chancellor, Domain wall encoding of discrete variables for quantum annealing and QAOA, Quantum Science Technology 4, 2019.