Encoding Methods

Variable Encoding

As you should already know, QUBO models are comprised only of binary variables. So when we are reformulating general optimization problems, one important step is to encode variables into binary ones.

ToQUBO currently implements 6 encoding techniques. Each method introduces a different number of variables, quadratic terms and linear terms. Also, they differ in the range of their coefficients($\Delta$).

Encoding	Binary Variables	# Linear terms	# Quadratic terms	$\Delta$

Binary	$O(log \ n)$	$O(log \ n)$	-	$O(n)$

Unary	$O(n)$	$O(n)$	-	$O(1)$

One-Hot	$O(n)$	$O(n)$	$O(n^2)$	$O(n)$

Domain-Wall	$O(n)$	$O(n)$	$O(n)$	$O(n)$

Bounded-Coefficient	$O(n)$	$O(n)$	-	$O(1)$

Arithmetic Prog	$O(\sqrt{n})$	$O(\sqrt{n})$	-	$O(\sqrt{n})$

Linear Encoding Methods

ToQUBO.Mirror — Type

Mirror

Mirrors binary variable $x \in \mathbb{B}$ with a twin variable $y \in \mathbb{B}$.

ToQUBO.LinearEncoding — Type

LinearEncoding

Every linear encoding $\xi$ is of the form

\[\xi(\mathbf{y}) = \alpha + \sum_{i = 1}^{n} \gamma_{i} y_{i}\]

ToQUBO.Linear — Type

Linear

ToQUBO.Binary — Type

Binary

Binary Expansion within the closed interval $[\alpha, \beta]$.

For a given variable $x \in [\alpha, \beta]$ we approximate it by

\[x \approx \alpha + \frac{(\beta - \alpha)}{2^{n} - 1} \sum_{i=0}^{n-1} {2^{i}\, y_i}\]

where $n$ is the number of bits and $y_i \in \mathbb{B}$.

ToQUBO.Unary — Type

Unary()

Let $x \in [a, b] \subset \mathbb{Z}, n = b - a, \mathbf{y} \in \mathbb{B}^{n}$.

\[x \approx \xi(\mathbf{y}) = a + \sum_{j = 1}^{b - a} y_{j}\]

ToQUBO.Arithmetic — Type

Arithmetic

ToQUBO.OneHot — Type

OneHot()

The one-hot encoding is a linear technique used to represent a variable $x \in \left\lbrace{\gamma_{j}}}\right\rbrace_{j \in \left[n\right]$.

The encoding function is combined with a constraint assuring that only one and exactly one of the expansion's variables $y_{j}$ is activated at a time.

\[\begin{array}{rl} x = \xi(\mathbf{y}) = & \sum_{j = 1}^{n} \gamma_{j} y_{j} \\ \mathrm{s.t.} & \sum_{j = 1}^{n} y_{j} = 1 \end{array}\]

Sequential Encoding Methods

Missing docstring.

Missing docstring for ToQUBO.SequentialEncoding. Check Documenter's build log for details.

ToQUBO.DomainWall — Type

DomainWall()

The Domain Wall[Chancellor2019] encoding method is a sequential approach that requires only $n - 1$ bits to represent $n$ distinct values.

Encoding Analysis

bits	linear	quadratic	$\Delta$

Domain Wall	$n - 1$	$n$		$O(n)$

Bounded Coefficients

ToQUBO.Bounded — Type

Bounded{Binary,T}(μ::T) where {T}

Let $x \in [a, b] \subset \mathbb{Z}$ and $n = b - a$.

First,

\[\begin{align*} 2^{k - 1} &\le \mu \\ \implies k &= \left\lfloor\log_{2} \mu + 1 \right\rfloor \end{align*}\]

Since

\[\sum_{j = 1}^{k} 2^{j - 1} = \sum_{j = 0}^{k - 1} 2^{j} = 2^{k} - 1\]

then, for $r \in \mathbb{N}$

\[n = 2^{k} - 1 + r \times \mu + \epsilon \implies r = \left\lfloor \frac{n - 2^{k} + 1}{\mu} \right\rfloor\]

and

\[\epsilon = n - 2^{k} + 1 - r \times \mu\]

\gamma_{j} = \left\lbrace\begin{array}{cl} 2^{j} & \text{if } 1 \le j \le k \ \mu & \text{if } k < j < r + k \ n - 2^k + 1 - r \times \mu & \text{otherwise} \end{array}\right. ```

Bounded{Unary,T}(μ::T) where {T}

Constraint Encoding

As you should already know, a QUBO model is unconstrained. So when ToQUBO is reformulating a problem, it needs to encode all constraints into the objective function loosing as little information as possible.

As constraints are introduced into the objective function, we need to make sure that they won't be violated. In order to do that, ToQUBO multiplies the encoded constraint by a large penalty $\rho$, so that any violation would result in an infeasible solution to the problem.

Sometimes, moving a constraint to the objective fuction might introduce higher-order terms (degree > 2). If that is the case, ToQUBO needs to reduce it back to a quadratic function.

As of today, ToQUBO provides encoding for the following constraints:

ToQUBO.toqubo_constraint — Function

toqubo_constraint

Returns the pseudo-boolean function associated to a given constraint from the source model.

Chancellor2019Nicholas Chancellor, Domain wall encoding of discrete variables for quantum annealing and QAOA, Quantum Science Technology 4, 2019.