Virtual Mapping

During reformulation, ToQUBO holds two distinct models, namely the Source Model and the Target Model. The source model is a generic MOI model restricted to the supported constraints. The target one is on the QUBO form used during the solving process. Both lie within a Virtual Model, which provides the necessary API integration and keeps all variable and constraint mapping tied together.

This is done in a transparent fashion for both agents since the user will mostly interact with the presented model, and the solvers will only access the generated one.

Virtual Model

ToQUBO.VirtualModel — Type

VirtualModel{T}(optimizer::Union{Nothing, Type{<:MOI.AbstractOptimizer}} = nothing) where {T}

This Virtual Model links the final QUBO formulation to the original one, allowing variable value retrieving and other features.

Virtual Variables

Every virtual model stores a collection of virtual variables, intended to provide a link between those in the source and those to be created in the target model. Each virtual variable stores encoding information for later expansion and evaluation.

ToQUBO.VirtualVariable — Type

Variable Expansion methods:

- Linear
- Unary
- Binary
- One Hot
- Domain Wall

References:

[1] Chancellor, N. (2019). Domain wall encoding of discrete variables for quantum annealing and QAOA. Quantum Science and Technology, 4(4), 045004. {doi}

ToQUBO.encode! — Function

encode!(model::VirtualModel{T}, v::VirtualVariable{T}) where {T}

Maps newly created virtual variable v within the virtual model structure. It follows these steps:

Maps v's source to it in the model's source mapping.

For every one of v's targets, maps it to itself and adds a binary constraint to it.

Adds v to the end of the model's varvec.