Compiler Settings

ToQUBO.Attributes.StableCompilation — Type

StableCompilation()

When set, this boolean flag enables stable reformulation methods, thus yielding predictable results.

Compiler Messages

ToQUBO.Attributes.Warnings — Type

Warnings()

Compiler Optimization

ToQUBO.Attributes.Optimization — Type

Optimization()

Working with target architectures

ToQUBO.Attributes.Architecture — Type

Architecture()

Selects which solver architecture to use. Defaults to QUBOTools.GenericArchitecture.

Quadratization

ToQUBO.Attributes.Quadratize — Type

Quadratize()

Boolean flag to conditionally perform the quadratization step. Is automatically set by the compiler when high-order functions are generated.

ToQUBO.Attributes.QuadratizationMethod — Type

QuadratizationMethod()

Defines which quadratization method to use. Available options are defined in the PBO submodule.

ToQUBO.Attributes.StableQuadratization — Type

StableQuadratization()

When set, this boolean flag enables stable quadratization methods, thus yielding predictable results. This is intended to be used during tests or other situations where deterministic output is desired. On the other hand, usage in production is not recommended since it requires increased memory and processing resources.

Variable & Constraint Encoding

ToQUBO.Attributes.VariableEncodingBits — Type

VariableEncodingBits()

ToQUBO.Attributes.DefaultVariableEncodingBits — Type

DefaultVariableEncodingBits()

ToQUBO.Attributes.VariableEncodingATol — Type

VariableEncodingATol()

ToQUBO.Attributes.DefaultVariableEncodingATol — Type

DefaultVariableEncodingATol()

Fallback value for VariableEncodingATol.

ToQUBO.Attributes.VariableEncodingMethod — Type

VariableEncodingMethod()

Available methods are:

Encoding.Binary (default)

Encoding.Unary

Encoding.Arithmetic

Encoding.OneHot

Encoding.DomainWall

Encoding.Bounded

The Encoding.Binary, Encoding.Unary and Encoding.Arithmetic encodings can have their expansion coefficients bounded by wrapping them with the Encoding.Bounded method.

ToQUBO.Attributes.DefaultVariableEncodingMethod — Type

DefaultVariableEncodingMethod()

Fallback value for VariableEncodingMethod.

ToQUBO.Attributes.VariableEncodingPenalty — Type

VariableEncodingPenalty()

Allows the user to set and retrieve the coefficients used for encoding variables when additional constraints are involved.

ToQUBO.Attributes.ConstraintEncodingPenalty — Type

ConstraintEncodingPenalty()

Allows the user to retrieve the coefficients used for encoding constraints.

Discretization

ToQUBO.Attributes.Discretize — Type

Discretize()

When set, this boolean flag guarantees that every coefficient in the final formulation is an integer.